requirements.txt
pyrogram
tgcrypto
imap-tools
python-dotenv
resend
